More Fixes

+++ AstroClient/AstroLibrary/Menu/AstroButtonAPI/QMTabButton.cs
@@ -7,37 +7,37 @@
 
     internal class QMTabButton : QMButtonBase
     {
-        internal QMTabButton(int Index, System.Action btnAction, String btnToolTip, Color? btnBackgroundColor = null, string LoadSprite = "")
+        internal QMTabButton(float Index, System.Action btnAction, String btnToolTip, Color? btnBackgroundColor = null, string LoadSprite = "")
         {
             initButton(Index, btnAction, btnToolTip, btnBackgroundColor, LoadSprite);
         }
-        internal QMTabButton(int Index, System.Action btnAction, String btnToolTip, Color? btnBackgroundColor = null, byte[] LoadSprite = null)
+        internal QMTabButton(float Index, System.Action btnAction, String btnToolTip, Color? btnBackgroundColor = null, byte[] LoadSprite = null)
         {
             initButton(Index, btnAction, btnToolTip, btnBackgroundColor, LoadSprite);
         }
 
-        internal void initButton(int Index, System.Action btnAction, String btnToolTip, Color? btnBackgroundColor = null, byte[] LoadSprite = null)
+        internal void initButton(float Index, System.Action btnAction, String btnToolTip, Color? btnBackgroundColor = null, byte[] LoadSprite = null)
         {
             btnType = "_QMTabButton_";
             button = UnityEngine.Object.Instantiate(QuickMenuStuff.TabButtonTemplate(), QuickMenuStuff.TabButtonTemplate().transform.parent, true);
             button.name = QMButtonAPI.identifier + btnType + Index;
             SetToolTip(btnToolTip);
             setAction(btnAction);
-            button.GetComponentInChildren<RectTransform>().SetSiblingIndex(Index);
+            button.GetComponentInChildren<RectTransform>().SetSiblingIndex((int)Index);
 
             if (LoadSprite != null)
                 button.LoadSprite(LoadSprite, "Icon");
 
             SetActive(true);
         }
-        internal void initButton(int Index, System.Action btnAction, String btnToolTip, Color? btnBackgroundColor = null, string LoadSprite = "")
+        internal void initButton(float Index, System.Action btnAction, String btnToolTip, Color? btnBackgroundColor = null, string LoadSprite = "")
         {
             btnType = "_QMTabButton_";
             button = UnityEngine.Object.Instantiate(QuickMenuStuff.TabButtonTemplate(), QuickMenuStuff.TabButtonTemplate().transform.parent, true);
             button.name = QMButtonAPI.identifier + btnType + Index;
             SetToolTip(btnToolTip);
             setAction(btnAction);
-            button.GetComponentInChildren<RectTransform>().SetSiblingIndex(Index);
+            button.GetComponentInChildren<RectTransform>().SetSiblingIndex((int)Index);
 
             if (LoadSprite != "")
                 button.LoadSprite(LoadSprite, "Icon");
